--- app/controllers/admin/production-deployment-how-to.md
@@ -1,22 +1,0 @@
-- login to `AWS`
-- go to `ECR` (https://ap-southeast-2.console.aws.amazon.com/ecr/repositories?region=ap-southeast-2) 
-- click repository name `pancake`
-- determine `image tag` you wish to deploy e.g `prod-1fsdf23`
-- go to `ECS`
-- click `Task Definitions` on left panel
-- select `pancake-task-defintion` checkbox
-- click `Create new revision`
-- scroll down to `Container Definitions`
-- click `Container Name`
-- change `364010606639.dkr.ecr.ap-southeast-2.amazonaws.com/pancake:[image-tag]` to use your `image tag`
-- click `Update`
-- click `Create`
-- select `Clusters` on left menu below heading `Amazon ECS`
-- click `pancake-cluster`
-- select `pancake-cluster-service-lb` checkbox
-- click `Update`
-- select latest revision (`[highest number](latest)`)
-- select force new deployment `checkbox`
-- click `Skip to review`
-- click `update service`
-- check `https://ap-southeast-2.console.aws.amazon.com/cloudwatch/home?region=ap-southeast-2#logs:` for updates
