+++ app/controllers/admin/rebate_forms_controller.rb
@@ -49,7 +49,7 @@ class Admin::RebateFormsController < Admin::BaseController
 
   # PATCH/PUT /admin/rebate_forms/1
   def update
-    @rebate_form = RebateFormsService.new(rebate_form_fields_params).update!
+    @rebate_form = RebateFormsUpdateService.new(rebate_form_fields_params).update!
     @rebate_form.update(updated_by: current_user.id)
     respond_with @rebate_form, location: admin_rebate_form_url(@rebate_form), notice: 'Rebate form was successfully updated.'
   end
@@ -72,7 +72,12 @@ class Admin::RebateFormsController < Admin::BaseController
   end
 
   def rebate_form_fields_params
-    params.permit(:id, :valuation_id, :total_rates, :location, :council, fields: {})
+    params.permit(:id,
+                  :valuation_id,
+                  :total_rates,
+                  :location,
+                  :council,
+                  fields: {})
   end
 
   def rebate_form_params
