+++ app/controllers/rebate_forms_controller.rb
@@ -25,7 +25,7 @@ class RebateFormsController < ApiController
   end
 
   def create
-    rebate_form = RebateFormsService.new(rebate_form_params).update!
+    rebate_form = RebateFormsService.new(rebate_form_params).create!
 
     render_jsonapi(rebate_form, scope: false)
   rescue RebateFormsService::Error
@@ -39,10 +39,8 @@ class RebateFormsController < ApiController
       .require(:api)
       .require(:data)
       .require(:attributes)
-      .permit(:id,
-              :valuation_id,
-              :council,
-              :total_rates,
+      .permit(:valuation_id,
+              :council_id,
               :location,
               fields: {},
               applicant_signature: {},
