+++ app/services/rates_importer_service.rb
@@ -33,8 +33,8 @@ class RatesImporterService
     rates_bill = RatesBill.find_by(property: property, rating_year: rating_year)
 
     if rates_bill.present?
-      current_rates = rates_bill.total_rates.to_f.round(2)
-      new_rates = (total_rates.to_f + total_water_rates.to_f).round(2)
+      # current_rates = rates_bill.total_rates.to_f.round(2)
+      # new_rates = (total_rates.to_f + total_water_rates.to_f).round(2)
       # raise 'mismatch total_rates' unless current_rates == new_rates
     else
       RatesBill.create!(
