+++ app/services/rebate_forms_service.rb
@@ -8,66 +8,35 @@ class RebateFormsService
     @rebate_form_attributes = rebate_form_attributes
   end
 
-  def update!
+  def create!
     council = find_council!
-    property = create_or_update_property!(council)
-    rebate_form = create_or_update_rebate_form!(property)
-    create_or_update_rates_bill!(property)
-    rebate_form.calc_rebate_amount!
-    raise Error if rebate_form.rebate.blank?
+    property = find_or_create_property!(council)
+    rebate_form = create_rebate_form!(property)
 
     rebate_form
   end
 
   private
 
-  def create_or_update_rebate_form!(property)
-    return update_rebate_form!(property) if @rebate_form_attributes['id']
-
-    create_rebate_form!(property)
-  end
-
-  def update_rebate_form!(property)
-    rebate_form = RebateForm.find_by(id: @rebate_form_attributes['id'])
-    property = rebate_form.property if property.id.nil?
-    remove_signatures_if_signed(rebate_form)
-    fields_to_update = rebate_form.fields.merge(fields_to_merge)
-    rebate_form.update!(property: property, valuation_id: property.valuation_id, fields: fields_to_update)
-    rebate_form
-  end
-
-  def remove_signatures_if_signed(rebate_form)
-    rebate_form.signatures.destroy_all && rebate_form.transition_to_not_signed_state if rebate_form.signed_state?
-  end
-
-  def update_fields
-    @rebate_form_attributes['fields']
-  end
-
-  def fields_to_merge
-    return update_fields unless update_fields.nil?
-
-    {}
-  end
-
   def create_rebate_form!(property)
     RebateForm.create!(property: property,
                        valuation_id: property.valuation_id,
+                       location: property.location,
+                       total_rates: total_rates(property),
                        rebate: 0,
                        fields: @rebate_form_attributes['fields'])
   end
 
-  def create_or_update_property!(council)
-    return create_or_update_property_with_valuation_id!(council) if @rebate_form_attributes['valuation_id']
+  def find_or_create_property!(council)
+    return find_or_create_property_with_valuation_id!(council) if @rebate_form_attributes['valuation_id']
 
     find_or_create_property_with_no_valuation_id!(council)
   end
 
-  def create_or_update_property_with_valuation_id!(council)
+  def find_or_create_property_with_valuation_id!(council)
     property = Property.find_or_create_by(valuation_id: @rebate_form_attributes['valuation_id'],
                                           rating_year: ENV['YEAR'],
                                           council: council)
-    property.update!(location: @rebate_form_attributes['location'])
     property
   end
 
@@ -84,11 +53,11 @@ class RebateFormsService
   def create_or_update_rates_bill!(property)
     rates_bill = RatesBill.find_or_create_by(property: property,
                                              rating_year: property.rating_year)
-    rates_bill.update!(total_rates: total_rates)
+    rates_bill.update!(total_rates: total_rates(property))
     rates_bill
   end
 
-  def total_rates
-    @rebate_form_attributes['total_rates'] || 0
+  def total_rates(property)
+    property.rates_bills[0]&.total_rates || 0
   end
 end
